+++ src/model.h
@@ -16,7 +16,7 @@ class Grid;
 
 class Model {
     private:
-        int k;
+        double k;
         double r;
         double pct_migrants;
         int leap_distance;
@@ -27,7 +27,7 @@ class Model {
 
     public:
         std::vector<Date> archaeo_dates;
-        Model(int start_date, int k, double r, double pct_migrants,
+        Model(int start_date, double k, double r, double pct_migrants,
               int leap_distance, double forest_threshold);
         ~Model();
         void init_pop();
